+++ genomepy/__init__.py
@@ -5,20 +5,18 @@ from loguru import logger
 
 from genomepy.__about__ import __author__, __version__
 from genomepy.annotation import Annotation
+from genomepy.config import manage_config
 from genomepy.exceptions import GenomeDownloadError
 from genomepy.functions import (
     clean,
     install_genome,
     list_available_genomes,
     list_installed_genomes,
-    manage_config,
-    manage_plugins,
 )
 from genomepy.genome import Genome
-from genomepy.provider import Provider
-
-list_available_providers = Provider.list_providers
-search = Provider.search_all
+from genomepy.plugins import manage_plugins
+from genomepy.providers import list_providers
+from genomepy.providers import search_all as search
 
 # logger is a singleton, configuration here will be used module-wide
 logger.remove()
